Extract bug monitor into dedicated crate
