+++ CHANGELOG.md
@@ -108,6 +108,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Began TAN-205 crate-boundary cleanup by relocating the provider auth
   credential store implementation into `tandem-providers` while keeping
   `tandem-core` compatibility re-exports for existing callers.
+- Continued TAN-205 crate-boundary cleanup by moving Bug Monitor domain types,
+  log parsing/evidence rendering, recurrence summaries, error provenance, and
+  GitHub publish logic into a new `tandem-bug-monitor` crate. `tandem-server`
+  keeps compatibility re-exports plus the AppState/HTTP/MCP host shim.
 - Replaced divergent tool-name normalization in parser, registry resolution,
   and approval classification with a shared `tandem-types` canonicalizer, plus
   a structured function-style invocation scanner and 30+ case parser corpus.
