+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "crates/tandem-wire",
     "crates/tandem-core",
     "crates/tandem-document",
+    "crates/tandem-bug-monitor",
     "crates/tandem-providers",
     "crates/tandem-tools",
     "crates/tandem-automation",
