+++ crates/tandem-bug-monitor/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "tandem-bug-monitor"
+version = "0.6.1"
+description = "Bug Monitor domain logic for Tandem"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/frumu-ai/tandem"
+edition = "2021"
+
+[dependencies]
+anyhow = "1"
+async-trait = "0.1"
+chrono = "0.4"
+regex = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sha2 = "0.10"
+tokio = { version = "1", features = ["fs", "process", "time"] }
+tracing = "0.1"
+uuid = { version = "1", features = ["v4"] }
+tandem-core = { path = "../tandem-core", version = "0.6.1" }
+tandem-types = { path = "../tandem-types", version = "0.6.1" }
+
+[dev-dependencies]
+tempfile = "3"
