+++ crates/tandem-server/Cargo.toml
@@ -52,6 +52,7 @@ tandem-providers = { path = "../tandem-providers", version = "0.6.1" }
 tandem-runtime = { path = "../tandem-runtime", version = "0.6.1" }
 tandem-tools = { path = "../tandem-tools", version = "0.6.1" }
 tandem-automation = { path = "../tandem-automation", version = "0.6.1" }
+tandem-bug-monitor = { path = "../tandem-bug-monitor", version = "0.6.1" }
 tandem-skills = { path = "../tandem-skills", version = "0.6.1" }
 tandem-memory = { path = "../tandem-memory", version = "0.6.1" }
 tandem-orchestrator = { path = "../tandem-orchestrator", version = "0.6.1" }
