+++ crates/tandem-server/src/bug_monitor/mod.rs
@@ -1,8 +1,5 @@
-pub mod comment_summary;
-pub mod error_provenance;
+pub use tandem_bug_monitor::{comment_summary, error_provenance, log_parser, types};
 pub mod log_artifacts;
-pub mod log_parser;
 pub mod log_watcher;
 pub mod regression_fixture;
 pub mod service;
-pub mod types;
